requirements.txt
meteostat==1.6.8
pandas==2.2.3
psycopg2_binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.35